+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   * [VM] Adding configuration key `AZK_DOCKER_MONITOR` to enable or disable docker monitor. By default is `true` for Linux and `false` if is not Linux;
   * [Cli] Send env TERM if is interactive terminal;
   * [Cli] Added the command `azk open [system]` which opens a system URL in browser #528;
+  * [Cli] Adding `deploy` command to cli;
+  * [Kernel] Changing system to get `command` from image, but respect this priority for `azk shell`: shell option (`--shell`) > Azkfile.js > docker image #534;
 
 ## v0.15.0 - (2015-09-15)
 
@@ -334,7 +336,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   * [Manifest] Adding support `docker_extra`
   * [Manifest] Adding support to 'disable' value in ports.
   * [Kernel] Now `azk` supports Linux \o/
-  * [Kenrel] You can now use the AZK to test and develop the AZK for Linux (see Azkfile.js).
+  * [Kernel] You can now use the AZK to test and develop the AZK for Linux (see Azkfile.js).
   * [Cmds] Now supports the `ssh escape sequence` in `shell` command.
 
 * Deprecations